Move the files around to create a better project structure
