+++ src/main.rs
@@ -8,19 +8,13 @@ Requires the following features: `cargo run --example system_tray --features "tr
 extern crate native_windows_gui as nwg;
 
 use nwg::NativeUi;
-use view::SystemTray;
+use crate::platform::launch_context;
+use crate::ui::view::SystemTray;
 
-mod view;
-mod mover;
-mod controller;
-mod delay_dialog;
-mod distance_dialog;
-mod settings;
-mod monitor_info;
-mod settings_path;
-mod autostart;
-mod launch_context;
-mod url_launcher;
+mod app;
+mod ui;
+mod platform;
+mod config;
 
 
 fn main() {
