+++ src/ui/dialogs/distance_dialog.rs
@@ -1,40 +1,52 @@
+use crate::platform::mover;
 use std::{thread, cell::RefCell};
-use std::time::Duration;
 use nwg::{ControlHandle, NativeUi, NumberSelectData};
-use crate::settings::{LOWEST_DELAY, MAX_DELAY};
+use crate::config::settings::LOWEST_MAX_DISTANCE;
 
-pub enum DelayDialogData {
+pub enum DistanceDialogData {
     Cancel,
-    Value(Duration),
+    Value(i32, i32),
 }
 
 #[derive(Default)]
-pub struct DelayDialog {
+pub struct DistanceDialog {
     window: nwg::Window,
     icon: nwg::Icon,
-    label: nwg::Label,
-    number_select: nwg::NumberSelect,
-    data: RefCell<Option<DelayDialogData>>,
+    label_x: nwg::Label,
+    label_y: nwg::Label,
+    number_select_x: nwg::NumberSelect,
+    number_select_y: nwg::NumberSelect,
+    data: RefCell<Option<DistanceDialogData>>,
     ok_button: nwg::Button,
     cancel_button: nwg::Button,
 }
 
-impl DelayDialog {
+impl DistanceDialog {
 
     /// Create the dialog UI on a new thread. The dialog result will be returned by the thread handle.
     /// To alert the main GUI that the dialog completed, this function takes a notice sender object.
-    pub(crate) fn popup(sender: nwg::NoticeSender, current: Duration) -> thread::JoinHandle<DelayDialogData> {
+    pub(crate) fn popup(sender: nwg::NoticeSender, current_value_x: i32, current_value_y: i32) -> thread::JoinHandle<DistanceDialogData> {
         return thread::spawn(move || {
             // Create the UI just like in the main function
-            let app = DelayDialog::build_ui(Default::default()).expect("Failed to build UI");
+            let app = DistanceDialog::build_ui(Default::default()).expect("Failed to build UI");
 
-            let number_select_data = NumberSelectData::Int {
-                value: current.as_secs() as i64,
-                step: 1, // 1 second steps
-                max: MAX_DELAY.as_secs() as i64,
-                min: LOWEST_DELAY.as_secs() as i64,
+            let (smallest_x, smallest_y) = mover::get_smallest_screen_size().unwrap_or((400, 400));
+
+            let number_select_data_x = NumberSelectData::Int {
+                value: current_value_x as i64,
+                step: 1,
+                max: smallest_x as i64 / 4,
+                min: LOWEST_MAX_DISTANCE as i64,
             };
-            app.number_select.set_data(number_select_data);
+            app.number_select_x.set_data(number_select_data_x);
+
+            let number_select_data_y = NumberSelectData::Int {
+                value: current_value_y as i64,
+                step: 1,
+                max: smallest_y as i64 / 4,
+                min: LOWEST_MAX_DISTANCE as i64,
+            };
+            app.number_select_y.set_data(number_select_data_y);
 
             nwg::dispatch_thread_events();
 
@@ -42,23 +54,28 @@ impl DelayDialog {
             sender.notice();
 
             // Return the dialog data
-            return app.data.take().unwrap_or(DelayDialogData::Cancel)
+            return app.data.take().unwrap_or(DistanceDialogData::Cancel)
         })
     }
 
     fn choose(&self, btn: &ControlHandle) {
         let mut data = self.data.borrow_mut();
         if btn == &self.ok_button {
-            let value = self.number_select.data();
-            if let Ok(parsed_value) = value.formatted_value().parse::<u64>() {
-                *data = Some(DelayDialogData::Value(Duration::from_secs(parsed_value)));
+            let value_x = self.number_select_x.data();
+            let value_y = self.number_select_y.data();
+
+            let value_x = value_x.formatted_value().parse::<i32>();
+            let value_y = value_y.formatted_value().parse::<i32>();
+
+            if value_x.is_ok() && value_y.is_ok() {
+                *data = Some(DistanceDialogData::Value(value_x.unwrap(), value_y.unwrap()));
             } else {
                 // TODO: Handle the error, if any
                 println!("Failed to parse value!");
-                *data = Some(DelayDialogData::Cancel);
+                *data = Some(DistanceDialogData::Cancel);
             }
         } else if btn == &self.cancel_button {
-            *data = Some(DelayDialogData::Cancel);
+            *data = Some(DistanceDialogData::Cancel);
         }
 
         self.window.close();
@@ -75,15 +92,15 @@ mod number_select_app_ui {
     use std::rc::Rc;
     use std::cell::RefCell;
     use std::ops::Deref;
-    use crate::view::ICON;
+    use crate::ui::view::ICON;
 
-    pub struct SpinDialogUI {
-        inner: Rc<DelayDialog>,
+    pub struct DistanceDialogUI {
+        inner: Rc<DistanceDialog>,
         default_handler: RefCell<Vec<nwg::EventHandler>>,
     }
 
-    impl NativeUi<SpinDialogUI> for DelayDialog {
-        fn build_ui(mut data: DelayDialog) -> Result<SpinDialogUI, nwg::NwgError> {
+    impl NativeUi<DistanceDialogUI> for DistanceDialog {
+        fn build_ui(mut data: DistanceDialog) -> Result<DistanceDialogUI, nwg::NwgError> {
             // Resources
             nwg::Icon::builder()
                 .source_bin(Option::from(ICON))
@@ -91,9 +108,9 @@ mod number_select_app_ui {
 
             // Controls
             nwg::Window::builder()
-                .size((320, 70))
+                .size((320, 100))
                 .center(true)
-                .title("Delay Select Dialog")
+                .title("Distance Dialog")
                 .icon(Some(&data.icon))
                 .flags(nwg::WindowFlags::WINDOW | nwg::WindowFlags::VISIBLE | nwg::WindowFlags::POPUP)
                 .build(&mut data.window)?;
@@ -105,17 +122,26 @@ mod number_select_app_ui {
                 .build(&mut grid)?;
 
             nwg::Label::builder()
-                .text("Value, in seconds:")
+                .text("Max distance x (pixels):")
                 .parent(&data.window)
-                .build(&mut data.label)?;
+                .build(&mut data.label_x)?;
+
+            nwg::Label::builder()
+                .text("Max distance y (pixels):")
+                .parent(&data.window)
+                .build(&mut data.label_y)?;
 
             nwg::NumberSelect::builder()
                 .size((152, 27))
                 .decimals( 0)
-                .min_int(200)
-                .value_int(30000)
                 .parent(&data.window)
-                .build(&mut data.number_select)?;
+                .build(&mut data.number_select_x)?;
+
+            nwg::NumberSelect::builder()
+                .size((152, 27))
+                .decimals( 0)
+                .parent(&data.window)
+                .build(&mut data.number_select_y)?;
 
             nwg::Button::builder()
                 .text("Ok")
@@ -127,13 +153,15 @@ mod number_select_app_ui {
                 .parent(&data.window)
                 .build(&mut data.cancel_button)?;
 
-            grid.add_child(0, 0, &data.label);
-            grid.add_child(1, 0, &data.number_select);
-            grid.add_child(0, 1, &data.ok_button);
-            grid.add_child(1, 1, &data.cancel_button);
+            grid.add_child(0, 0, &data.label_x);
+            grid.add_child(1, 0, &data.number_select_x);
+            grid.add_child(0, 1, &data.label_y);
+            grid.add_child(1, 1, &data.number_select_y);
+            grid.add_child(0, 2, &data.ok_button);
+            grid.add_child(1, 2, &data.cancel_button);
 
             // Wrap-up
-            let ui = SpinDialogUI {
+            let ui = DistanceDialogUI {
                 inner: Rc::new(data),
                 default_handler: Default::default(),
             };
@@ -147,12 +175,12 @@ mod number_select_app_ui {
                     match evt {
                         E::OnButtonClick => {
                             if &handle == &ui.ok_button || &handle == &ui.cancel_button {
-                                DelayDialog::choose(&ui, &handle);
+                                DistanceDialog::choose(&ui, &handle);
                             }
                         }
                         E::OnWindowClose => {
                             if &handle == &ui.window {
-                                DelayDialog::exit(&ui);
+                                DistanceDialog::exit(&ui);
                             }
                         }
                         _ => {}
@@ -168,7 +196,7 @@ mod number_select_app_ui {
         }
     }
 
-    impl Drop for SpinDialogUI {
+    impl Drop for DistanceDialogUI {
         /// To make sure that everything is freed without issues, the default handler must be unbound.
         fn drop(&mut self) {
             let mut handlers = self.default_handler.borrow_mut();
@@ -178,10 +206,10 @@ mod number_select_app_ui {
         }
     }
 
-    impl Deref for SpinDialogUI {
-        type Target = DelayDialog;
+    impl Deref for DistanceDialogUI {
+        type Target = DistanceDialog;
 
-        fn deref(&self) -> &DelayDialog {
+        fn deref(&self) -> &DistanceDialog {
             &self.inner
         }
     }
