+++ src/ui/dialogs/mod.rs
@@ -1,0 +1,2 @@
+pub mod delay_dialog;
+pub mod distance_dialog;
