+++ src/ui/mod.rs
@@ -1,0 +1,2 @@
+pub mod view;
+pub mod dialogs;
